# openapi.yaml
openapi: 3.0.0
servers:
  - url: http://localhost8080
info:
  description: Contract API For The Skill-Matrix Application. This API contains all the neccessary endpoints for CRUD operations based on project requirements. In a nutshell, the application for which this API was created, contains all the active employees currently working for SoftwareONE. The purpose of this application is to aggregate all employees in one place so that any manager can access and contact any employee based on skills and other criterias.
  version: "1.0.0"
  title: Skill-Matrix API 
  contact:
    email: zourab.lavrentiadis@softwareone.com
  license:
    name: SoftwareONE
    url: 'http://www.softwareone.com'
...
paths:
...
  /department:
    get:
      tags:
        - department
      summary: fetch all departments from the database
      operationId: getDepartments
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Department'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server   
  
    post:
      tags:
        - department
      summary: add a department to the database
      operationId: addDepartment
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Department'
      responses:
        201:
           description: created
           content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server   
  /department/{departmentId}:
    get:
      tags:
        - department
      summary: find a department by id
      operationId: getDepartmentById
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Department'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server   
  
    put:
      tags:
        - department
      summary: update a department by id
      operationId: updateDepartment
      parameters:
      - in: path
        name: departmentId
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
               $ref: '#/components/schemas/Department'
      responses:
        201:
          description: item created
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server   
          
    delete:
      tags:
        - department
      summary: delete a department by id
      operationId: deleteDepartment
      parameters:
      - in: path
        name: departmentId
        required: true
        schema:
          type: integer
      responses:
        201:
          description: deleted
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server      


 #category
  /category:
    get:
      tags:
        - category
      summary: fetch all categories from the database
      operationId: getSkillCategory
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server      

    post:
      tags:
        - category
      summary: add a category to the database
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        201:
           description: created
           content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server    
  /category/{categoryId}:
    get:
      tags:
        - category
      summary: find category by id
      operationId: getCategoryById
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server    
   
    put:
      tags:
        - category
      summary: update a category by id
      operationId: updateCategory
      parameters:
      - in: path
        name: categoryId
        required: true
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
               $ref: '#/components/schemas/Category'
      responses:
        201:
          description: created
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server    
          
    delete:
      tags:
        - category
      summary: delete a category by id
      operationId: deleteCategory
      parameters:
      - in: path
        name: categoryId
        required: true
        schema:
          type: integer
      responses:
        201:
          description: deleted
        4XX:
          description: problem caused by the user
        5XX:
          description: problem caused by the server  


 #location
...
components:
  schemas:
...
    Department:
      type: object
      properties:
        departmentId:
          type: integer
          example: 1
        department:
          type: string
          enum:
            - application
            - sales
            - marketing  
    Category:
      type: object
      properties:
        categoryId:
          type: integer
          example: 1
        categoryName:
          type: string
          enum:
            - framework
            - cloud
            - programming
...